+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## (not yet)
+## v1.1.0 (2018-02-05)
 
 ### Breaking change: matter.js is split from dxopal.js